changes for authorization

+++ backend/routes/posts.js
@@ -42,7 +42,11 @@ router.post('', checkAuth, multer({ storage }).single('image'), async (req, res)
   try {
     const url = `${req.protocol}://${req.get('host')}`;
     const { title, content } = req.body;
-    const post = new Post({ title, content, imagePath: `${url}/images/${req.file.filename}` });
+    const post = new Post({
+       title,
+       content,
+       imagePath: `${url}/images/${req.file.filename}`,
+       creator: req.userData.userId });
     await post.save();
     const filteredPost = withOnly(['title', 'content'], post);
     res.status(201).json({ post: { id: post._id, ...filteredPost } });
@@ -84,8 +88,12 @@ router.get('/:id', checkAuth, async (req, res, next) => {
 
 router.delete('/:id', checkAuth, async (req, res, next) => {
   try {
-    const post = await Post.deleteOne({_id: req.params.id});
-    res.status(200).json({ post });
+    const post = await Post.deleteOne({_id: req.params.id, creator: req.userData.userId});
+    if (post.n > 0) {
+      res.status(200).json({ messsage: 'Delete successfull' });
+    } else {
+      res.status(401).json({ messsage: 'Not Authorized' });
+    }
   } catch (error) {
     console.log(error);
     res.status(500).json({ error: error.message });
@@ -100,8 +108,13 @@ router.put('/:id', checkAuth, multer({ storage }).single('image'), async (req, r
       const url = `${req.protocol}://${req.get('host')}`;
       path = `${url}/images/${req.file.filename}`;
     }
-    const post = await Post.updateOne( { _id: req.params.id }, { title, content, imagePath: path });
-    res.status(200).json( { post } );
+    const post = await Post.updateOne( { _id: req.params.id, creator: req.userData.userId }, { title, content, imagePath: path, creator: req.userData.userId });
+    console.log(post);
+    if (post.nModified > 0) {
+      res.status(200).json({ messsage: 'Update successfull' });
+    } else {
+      res.status(401).json({ messsage: 'Not Authorized' });
+    }
   } catch (error) {
     console.log(error);
     res.status(500).json({ error: error.message });
